FIX: Fixing PrintXML and Directory Structure
