+++ TheatricalPlayersRefactoringKata/Models/Invoice.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Models;
 
 public class Invoice
 {
@@ -13,8 +13,8 @@ public class Invoice
 
     public Invoice(string customer, List<Performance> performance)
     {
-        this._customer = customer;
-        this._performances = performance;
+        _customer = customer;
+        _performances = performance;
     }
